+++ src/Base.js
@@ -2,15 +2,13 @@
 
 import 'airbnb-js-shims';
 import 'lib/i18n';
-
 import 'assets/fonts/menomonia.css';
 import 'assets/fonts/opensans.css';
 
+import rootReducer from 'features/reducer';
 import { createStore, applyMiddleware, compose } from 'redux';
 import thunk from 'redux-thunk';
 import { Provider } from 'react-redux';
-
-import rootReducer from 'features/reducer';
 
 const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
 const store = createStore(rootReducer, composeEnhancers(